Split `Notification` "pure" model code from apps details
The `Notification` model is littered with functionality that has to do
with the view layer. To move `Notification` to WordPressData in a timely
fashion, we need to separate the model description from all that other
logic.
